allowing two implementations of gpio
